requirements.txt
keyboard==0.13.5
PyQt5==5.15.10
PyQt5_sip==12.13.0
PyYAML==6.0.1
PyAudio==0.2.14
SpeechRecognition==3.10.1
